remodelando o projeto

+++ portal/delete.php
@@ -1,22 +1,31 @@
 <?php
     require "../conexaoMysql.php";
     $pdo = mysqlConnect();
-
-    $idAnuncio = $_GET["id"];
     session_start();
-    $email = $_SESSION["email"];
-    $sql = <<<sql
-        SELECT idAnunciante FROM anunciante
-            WHERE email = ?
-    sql;
-    $stmt = $pdo->prepare($sql);
-    $stmt->execute([$email]);
-    $idAnunciante = $stmt->fetch();
+    $email = $_SESSION["email"] ?? "";
+    if($email === ""){
+        header("Location: ../index.html");
+        exit();
+    }
+    try{
+        $idAnuncio = $_GET["id"];
+        
+        $sql = <<<sql
+            SELECT idAnunciante FROM anunciante
+                WHERE email = ?
+        sql;
+        $stmt = $pdo->prepare($sql);
+        $stmt->execute([$email]);
+        $row = $stmt->fetch();
+        $idAnunciante = $row["idAnunciante"];
 
-    $sql = <<<sql
-        DELETE FROM anuncio
-            WHERE anuncio.idAnuncio = ? AND anuncio.idAnunciante = ?
-    sql;
-    $stmt = $pdo->prepare($sql);
-    $stmt->execute([$idAnuncio, $idAnunciante]);
+        $sql = <<<sql
+            DELETE FROM anuncio
+                WHERE anuncio.idAnuncio = ? AND anuncio.idAnunciante = ?
+        sql;
+        $stmt = $pdo->prepare($sql);
+        $stmt->execute([$idAnuncio, $idAnunciante]);
+    }catch (Exception $e) {
+        exit('Ocorreu uma falha: ' . $e->getMessage());
+    }
 ?>
